moved Leaderboard.js and chart components into subdirectory ./src/components/leaderboardview
