+++ frontend/src/components/leaderboardview/barchart_minutes_per_person.js
